Fixed add sponsor to a program endpoint bug. Added get pledge endpoint to sponsor
I switched the addSponsor controller function of Program to use Promise.all so that both updates to Program and Sponsor document will be required to resolve or it will return an error if one fails.

Further improvements to this add sponsor to program endpoint can be adding restrictions such as a) a sponsor must have sufficient balance to pledge and b) a program can't have any more sponsor if NOT at crowdfunding phase anymore

The get pledge endpoint assumes that a sponsor can pledge multiple times, hence why .map() was used. If the program specification were to say that a program can only pledge one time, .filter() can be used instead.